requirements.txt
fastapi
uvicorn
requests
langchain
langchain-openai
python-dotenv